\name{bandTable}
\alias{bandTable}
\docType{data}
\title{Human chromosomes information}
\description{
    bandTable contains chromosomes band information and centromer location. It 
    is used by mCSEAPlot() function to plot the chromosome track.
}
\usage{data(bandTable)}
\format{
    data.frame
}
\source{
    Obtained with Gviz's IdeogramTrack() function.
} 
\keyword{datasets}
