\name{exprTest}
\alias{exprTest}
\docType{data}
\title{Expression data example}
\description{
    exprTest is a subset of 100 genes' microarray expression data for 20 
    bone marrow samples: 10 from Acute Lymphoblastic Leukemia patients and 
    10 from healthy patients. It is useful to test mCSEAIntegrate function.
}
\usage{data(exprTest)}
\format{
    matrix
}
\source{
    Obtained from the leukemiasEset data package
} 
\keyword{datasets}
