% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCSEAPlot.R
\name{mCSEAPlot}
\alias{mCSEAPlot}
\title{Plot mCSEA results}
\usage{
mCSEAPlot(data, pheno, mCSEAResults, regionType, region, column = 1,
  platform = "450k", extend = 1000, chromosome = TRUE,
  leadingEdge = TRUE, CGI = FALSE, genes = TRUE,
  transcriptAnnotation = "transcript", makePDF = TRUE)
}
\arguments{
\item{data}{A data frame or a matrix containing Illumina's CpG probes in rows
and samples in columns}

\item{pheno}{A data frame or a matrix containing samples in rows and
covariates in columns}

\item{mCSEAResults}{The object generated by mCSEATest function}

\item{regionType}{The region type to be represented. Must be one of
"promoters", "genes", "CGI" or "custom"}

\item{region}{The region of interest to be represented (e.g. gene name, CGI
name...)}

\item{column}{The column name or number from pheno used to split the samples
into groups (first column is used by default)}

\item{platform}{Platform used to get the methylation data (450k or EPIC)}

\item{extend}{The number of base pairs to extend the plot around the region
of interest (default = 1000 bp)}

\item{chromosome}{If TRUE, represent the chromosome and genome axis}

\item{leadingEdge}{If TRUE, represent the bars indicating if each CpG belongs
to the mCSEA leading edge (green) or not (red)}

\item{CGI}{If TRUE, represent the annotated CpG islands}

\item{genes}{If TRUE, represent the annotated genes}

\item{transcriptAnnotation}{Labels showed at the genes track. Must be one of
"transcript" (default), "symbol", "gene", "exon" or "feature"}

\item{makePDF}{If TRUE, generate save the plot in pdf format in the working
directory. Otherwise, draw the plot in the active graphics window}
}
\value{
'NULL'
}
\description{
Generate a graphic with the genomic context of the selected region, showing
methylation status at each CpG site of different samples groups
}
\examples{
library(mCSEAdata)
data(mcseadata)
\dontrun{
myRank <- rankProbes(betaTest, phenoTest, refGroup = "Control")
myResults <- mCSEATest(myRank, regionsTypes = "promoters",
platform = "EPIC")
}
data(precomputedmCSEA)
mCSEAPlot(betaTest, phenoTest, myResults, "promoters", "CLIC6",
platform = "EPIC", transcriptAnnotation = "symbol", makePDF = FALSE)
}
\seealso{
\code{\link{rankProbes}}, \code{\link{mCSEATest}},
\code{\link{mCSEAPlotGSEA}}
}
\author{
Jordi Martorell Marugán, \email{jordi.martorell@genyo.es}
}
