\name{precomputedmCSEA}
\alias{precomputedmCSEA}
\alias{myRank}
\alias{myResults}
\docType{data}
\title{Precomputed mCSEA results}
\description{
    myRank is a result of rankProbes function.
    myResults is a result of mCSEATest function.
}
\usage{data(precomputedmCSEA)}
\format{
    vector (\code{myRank}) and list data.frame (\code{myResults})
}
\source{
    Both objects were obtained with the example commands in the mCSEA
    vignette.
} 
\keyword{datasets}
